+++ packages/ext/src/utils/document.ts
@@ -49,17 +49,25 @@ export async function getShadowRootAsync(): Promise<ShadowRoot> {
 
 export function injectUserSettingCssVariables(cssVariableText: string) {
   const validText = filterValidCssVariableText(cssVariableText)
-  getShadowRootAsync().then(root => {
-    const style = document.createElement('style');
-    style.innerHTML = `
+  injectCssIntoShadowRoot(`
 html, :host {
   ${validText}
+}`)
 }
-    `;
+
+export function injectCssIntoShadowRoot(css: string) {
+  getShadowRootAsync().then(root => {
+    const style = document.createElement('style');
+    style.innerHTML =css;
     root.querySelector('head')?.append(style)
   })
 }
 
+/**
+ * for filter user-custom css variables text, filter out line without valid prefix
+ * @param cssVariableText 
+ * @returns 
+ */
 export function filterValidCssVariableText(cssVariableText: string) {
   return cssVariableText
     .split('\n')
@@ -68,4 +76,49 @@ export function filterValidCssVariableText(cssVariableText: string) {
     .join('\n')
 }
 
+/**
+ * listener for SPA route change
+ * set a delay of 1000ms to trigger callback, because that on pathname changed, documents may have not been rendered completely, 
+ * 1000ms as a threshold will not 100% solve the problem, the complete time depends on Framework and network, 
+ * because there is noway to perfectly listen the SPA rerendere done, this is a compromise solution.
+ * @param callback 
+ */
+export function onSpaRouteChange(callback: Function) {
+  let pathname = window.location.pathname
 
+  function detectChange() {
+    let curPathname = window.location.pathname
+    if (curPathname !== pathname) {
+      pathname = curPathname
+      callback()
+    }
+  }
+
+
+  // // 监听 DOM 变化
+  const observer = new MutationObserver(() => {
+    /**
+     * 
+     */
+    setTimeout(() => {
+      detectChange()
+    }, 1000)
+  });
+
+  observer.observe(document.body, { childList: true, subtree: true });
+  return { disconnect: () => observer.disconnect() }
+}
+
+
+export async  function writeTextToClipboard(text: string) {
+  if(navigator.clipboard)
+    await navigator.clipboard.writeText(text)
+  else{
+    const textarea = document.createElement("textarea");
+    textarea.value = text;
+    document.body.appendChild(textarea);
+    textarea.select();
+    document.execCommand("copy");
+    document.body.removeChild(textarea);
+  }
+}